Added a dockerfile for building and testing the package (#195)
* Added a dockerfile for building and testing the package

This dockerfile can be used to setup and run the tests in the Python Deequ package. This way, we do not need to install any dependencies in our local workspaces. Right now, it only builds against Spark version 3.3. Will be adding other versions in a future PR.

Verified that the docker run output is the same as that of the PR workflow.

* Locked Poetry version to 1.7.1